fix: - renamed "LoadingIndicator" to "AppLoading"
